Ensure backward compatibility for attributes stored in NVS with primitive data type
Removed config option ESP_MATTER_NVS_USE_COMPACT_ATTR_STORAGE.

Previously, the ESP_MATTER_NVS_USE_COMPACT_ATTR_STORAGE option
determined whether to store attributes as a blob or a primitive data
type in NVS. Since this configuration is not backward compatible, we now
read the attribute as a primitive data type if not found. If it's a
blob, we read it and rewrite it as a primitive data type.