components/esp-matter: fix the inconsistency between feature flag and added features
After restructuring we were only adding features that are marked with
O.a or O.a+ conformance. But, it leads to the inconsistency where user
speicied features will not be added.

For clusters which has O.a or O.a+ feature conformance, linking all the
supported features in cluster create to address the incosistency.

Also, moved the feature create in the server part.